fix: remove extra allocs from alternate case impl
